blufor_facewear - move camo nets into separate includes, add headgear slot balaclavas
