+++ addons/blufor_facewear/CfgGlasses.hpp
@@ -1,136 +1,6 @@
 class cfgGlasses
 {
 	class None;
-	class tmtm_f_camoNetting_arid_1: None
-	{
-		author = "Trenchgun";
-		scope = 2;
-		scopeArsenal = 2;
-		displayName = "Camo Netting (Arid 1)";
-		model = "\A3\Characters_F\BLUFOR\headgear_b_helmet_camo";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_arid_ca.paa";
-		identityTypes[] =
-		{
-			"NoGlasses",0,"G_NATO_default",0,"G_NATO_casual",0,"G_NATO_pilot",0,"G_NATO_recon",50,"G_NATO_SF",0,"G_NATO_sniper",0,
-			"G_NATO_diver",0,"G_IRAN_default",0,"G_IRAN_diver",0,"G_GUERIL_default",0,"G_HAF_default",0,"G_CIVIL_female",0,"G_CIVIL_male",0
-		};
-		mass = 0.1;
-		hiddenSelections[] = {"camo1", "camo2"};
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_1_ca.paa"};
-	};
-	class tmtm_f_camoNetting_arid_2: tmtm_f_camoNetting_arid_1
-	{
-		displayName = "Camo Netting (Arid 2)";
-		scope = 2;
-		scopeArsenal = 2;
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_2_ca.paa"};
-	};
-	class tmtm_f_camoNetting_arid_3: tmtm_f_camoNetting_arid_1
-	{
-		displayName = "Camo Netting (Arid 3)";
-		scope = 2;
-		scopeArsenal = 2;
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_3_ca.paa"};
-	};
-	class tmtm_f_camoNetting_arid_4: tmtm_f_camoNetting_arid_1
-	{
-		displayName = "Camo Netting (Arid 4)";
-		scope = 2;
-		scopeArsenal = 2;
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_4_ca.paa"};
-	};
-	class tmtm_f_camoNetting_arid_5: tmtm_f_camoNetting_arid_1
-	{
-		displayName = "Camo Netting (Arid 5)";
-		scope = 2;
-		scopeArsenal = 2;
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_5_ca.paa"};
-	};
-	class tmtm_f_camoNetting_semiArid_1: None
-	{
-		author = "Trenchgun";
-		scope = 2;
-		scopeArsenal = 2;
-		displayName = "Camo Netting (Semi-Arid 1)";
-		model = "\A3\Characters_F\BLUFOR\headgear_b_helmet_camo";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_semiArid_ca.paa";
-		identityTypes[] =
-		{
-			"NoGlasses",0,"G_NATO_default",0,"G_NATO_casual",0,"G_NATO_pilot",0,"G_NATO_recon",50,"G_NATO_SF",0,"G_NATO_sniper",0,
-			"G_NATO_diver",0,"G_IRAN_default",0,"G_IRAN_diver",0,"G_GUERIL_default",0,"G_HAF_default",0,"G_CIVIL_female",0,"G_CIVIL_male",0
-		};
-		mass = 0.1;
-		hiddenSelections[] = {"camo1", "camo2"};
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_semiArid_1_ca.paa"};
-	};
-	class tmtm_f_camoNetting_lush_1: None
-	{
-		author = "Trenchgun";
-		scope = 2;
-		scopeArsenal = 2;
-		displayName = "Camo Netting (Lush 1)";
-		model = "\A3\Characters_F\BLUFOR\headgear_b_helmet_camo";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_lush_ca.paa";
-		identityTypes[] =
-		{
-			"NoGlasses",0,"G_NATO_default",0,"G_NATO_casual",0,"G_NATO_pilot",0,"G_NATO_recon",50,"G_NATO_SF",0,"G_NATO_sniper",0,
-			"G_NATO_diver",0,"G_IRAN_default",0,"G_IRAN_diver",0,"G_GUERIL_default",0,"G_HAF_default",0,"G_CIVIL_female",0,"G_CIVIL_male",0
-		};
-		mass = 0.1;
-		hiddenSelections[] = {"camo1", "camo2"};
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_lush_1_ca.paa"};
-	};
-	class tmtm_f_camoNetting_lush_2: tmtm_f_camoNetting_lush_1
-	{
-		displayName = "Camo Netting (Lush 2)";
-		scope = 2;
-		scopeArsenal = 2;
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_lush_2_ca.paa"};
-	};
-	class tmtm_f_camoNetting_jungle_1: None
-	{
-		author = "Trenchgun";
-		scope = 2;
-		scopeArsenal = 2;
-		displayName = "Camo Netting (Jungle 1)";
-		model = "\A3\Characters_F\BLUFOR\headgear_b_helmet_camo";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_jungle_ca.paa";
-		identityTypes[] =
-		{
-			"NoGlasses",0,"G_NATO_default",0,"G_NATO_casual",0,"G_NATO_pilot",0,"G_NATO_recon",50,"G_NATO_SF",0,"G_NATO_sniper",0,
-			"G_NATO_diver",0,"G_IRAN_default",0,"G_IRAN_diver",0,"G_GUERIL_default",0,"G_HAF_default",0,"G_CIVIL_female",0,"G_CIVIL_male",0
-		};
-		mass = 0.1;
-		hiddenSelections[] = {"camo1", "camo2"};
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_1_ca.paa"};
-	};
-	class tmtm_f_camoNetting_jungle_2: tmtm_f_camoNetting_jungle_1
-	{
-		displayName = "Camo Netting (Jungle 2)";
-		scope = 2;
-		scopeArsenal = 2;
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_2_ca.paa"};
-	};
-	class tmtm_f_camoNetting_jungle_3: tmtm_f_camoNetting_jungle_1
-	{
-		displayName = "Camo Netting (Jungle 3)";
-		scope = 2;
-		scopeArsenal = 2;
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_3_ca.paa"};
-	};
-	class tmtm_f_camoNetting_jungle_4: tmtm_f_camoNetting_jungle_1
-	{
-		displayName = "Camo Netting (Jungle 4)";
-		scope = 2;
-		scopeArsenal = 2;
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_4_ca.paa"};
-	};
-	class tmtm_f_camoNetting_jungle_5: tmtm_f_camoNetting_jungle_1
-	{
-		displayName = "Camo Netting (Jungle 5)";
-		scope = 2;
-		scopeArsenal = 2;
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_5_ca.paa"};
-	};
-	
+
+	#include "include/CfgGlasses.hpp"	
 };
