+++ addons/blufor_facewear/CfgWeapons.hpp
@@ -2,213 +2,11 @@ class CfgWeapons
 {
 	class NVGoggles;
 
-    class tmtm_nvg_camoNetting_base: NVGoggles
-    {
-        scope=0;
-
-		model = "\A3\Characters_F\BLUFOR\headgear_b_helmet_camo";
-		hiddenSelections[] = {"camo1", "camo2"};
-
-		modelOptics="\A3\Weapons_F\empty";
-		visionMode[]= {
-			"Normal",
-			"Normal"
-		};
-    };
-
-	class tmtm_nvg_camoNetting_arid_1: tmtm_nvg_camoNetting_base
+	class ItemCore;
+	class HelmetBase: ItemCore
 	{
-		scope=2;
-        scopeArsenal=2;
-        
-        author = "Trenchgun";
-
-		displayName = "Camo Netting (Arid 1)";
-        descriptionShort="Helmet Netting, Camo, Arid";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_arid_ca.paa";
-
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_1_ca.paa"};
-
-        TMTM_NVG_CAMONETTING_ITEMINFO("\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_1_ca.paa")
+		class ItemInfo;
 	};
-	class tmtm_nvg_camoNetting_arid_2: tmtm_nvg_camoNetting_base
-	{
-		scope=2;
-        scopeArsenal=2;
-        
-        author = "Trenchgun";
 
-		displayName = "Camo Netting (Arid 2)";
-        descriptionShort="Helmet Netting, Camo, Arid";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_arid_ca.paa";
-
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_2_ca.paa"};
-
-        TMTM_NVG_CAMONETTING_ITEMINFO("\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_2_ca.paa")
-	};
-	class tmtm_nvg_camoNetting_arid_3: tmtm_nvg_camoNetting_base
-	{
-		scope=2;
-        scopeArsenal=2;
-        
-        author = "Trenchgun";
-
-		displayName = "Camo Netting (Arid 3)";
-        descriptionShort="Helmet Netting, Camo, Arid";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_arid_ca.paa";
-
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_3_ca.paa"};
-
-        TMTM_NVG_CAMONETTING_ITEMINFO("\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_3_ca.paa")
-	};
-	class tmtm_nvg_camoNetting_arid_4: tmtm_nvg_camoNetting_base
-	{
-		scope=2;
-        scopeArsenal=2;
-        
-        author = "Trenchgun";
-
-		displayName = "Camo Netting (Arid 4)";
-        descriptionShort="Helmet Netting, Camo, Arid";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_arid_ca.paa";
-
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_4_ca.paa"};
-
-        TMTM_NVG_CAMONETTING_ITEMINFO("\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_4_ca.paa")
-	};
-	class tmtm_nvg_camoNetting_arid_5: tmtm_nvg_camoNetting_base
-	{
-		scope=2;
-        scopeArsenal=2;
-        
-        author = "Trenchgun";
-
-		displayName = "Camo Netting (Arid 5)";
-        descriptionShort="Helmet Netting, Camo, Arid";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_arid_ca.paa";
-
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_5_ca.paa"};
-
-        TMTM_NVG_CAMONETTING_ITEMINFO("\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_arid_5_ca.paa")
-	};
-	class tmtm_nvg_camoNetting_semiArid_1: tmtm_nvg_camoNetting_base
-	{
-		scope=2;
-        scopeArsenal=2;
-        
-        author = "Trenchgun";
-
-		displayName = "Camo Netting (Semi-Arid 1)";
-        descriptionShort="Helmet Netting, Camo, Semi-Arid";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_semiArid_ca.paa";
-
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_semiArid_1_ca.paa"};
-
-        TMTM_NVG_CAMONETTING_ITEMINFO("\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_semiArid_1_ca.paa")
-	};
-	class tmtm_nvg_camoNetting_lush_1: tmtm_nvg_camoNetting_base
-	{
-		scope=2;
-        scopeArsenal=2;
-        
-        author = "Trenchgun";
-
-		displayName = "Camo Netting (Lush 1)";
-        descriptionShort="Helmet Netting, Camo, Lush";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_lush_ca.paa";
-
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_lush_1_ca.paa"};
-
-        TMTM_NVG_CAMONETTING_ITEMINFO("\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_lush_1_ca.paa")
-	};
-	class tmtm_nvg_camoNetting_lush_2: tmtm_nvg_camoNetting_base
-	{
-		scope=2;
-        scopeArsenal=2;
-        
-        author = "Trenchgun";
-
-		displayName = "Camo Netting (Lush 2)";
-        descriptionShort="Helmet Netting, Camo, Lush";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_lush_ca.paa";
-
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_lush_2_ca.paa"};
-
-        TMTM_NVG_CAMONETTING_ITEMINFO("\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_lush_2_ca.paa")
-	};
-	class tmtm_nvg_camoNetting_jungle_1: tmtm_nvg_camoNetting_base
-	{
-		scope=2;
-        scopeArsenal=2;
-        
-        author = "Trenchgun";
-
-		displayName = "Camo Netting (Jungle 1)";
-        descriptionShort="Helmet Netting, Camo, Jungle";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_jungle_ca.paa";
-
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_1_ca.paa"};
-
-        TMTM_NVG_CAMONETTING_ITEMINFO("\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_1_ca.paa")
-	};
-	class tmtm_nvg_camoNetting_jungle_2: tmtm_nvg_camoNetting_base
-	{
-		scope=2;
-        scopeArsenal=2;
-        
-        author = "Trenchgun";
-
-		displayName = "Camo Netting (Jungle 2)";
-        descriptionShort="Helmet Netting, Camo, Jungle";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_jungle_ca.paa";
-
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_2_ca.paa"};
-
-        TMTM_NVG_CAMONETTING_ITEMINFO("\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_2_ca.paa")
-	};
-	class tmtm_nvg_camoNetting_jungle_3: tmtm_nvg_camoNetting_base
-	{
-		scope=2;
-        scopeArsenal=2;
-        
-        author = "Trenchgun";
-
-		displayName = "Camo Netting (Jungle 3)";
-        descriptionShort="Helmet Netting, Camo, Jungle";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_jungle_ca.paa";
-
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_3_ca.paa"};
-
-        TMTM_NVG_CAMONETTING_ITEMINFO("\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_3_ca.paa")
-	};
-	class tmtm_nvg_camoNetting_jungle_4: tmtm_nvg_camoNetting_base
-	{
-		scope=2;
-        scopeArsenal=2;
-        
-        author = "Trenchgun";
-
-		displayName = "Camo Netting (Jungle 4)";
-        descriptionShort="Helmet Netting, Camo, Jungle";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_jungle_ca.paa";
-
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_4_ca.paa"};
-
-        TMTM_NVG_CAMONETTING_ITEMINFO("\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_4_ca.paa")
-	};
-	class tmtm_nvg_camoNetting_jungle_5: tmtm_nvg_camoNetting_base
-	{
-		scope=2;
-        scopeArsenal=2;
-        
-        author = "Trenchgun";
-
-		displayName = "Camo Netting (Jungle 5)";
-        descriptionShort="Helmet Netting, Camo, Jungle";
-		picture = "\tmtm\pack\blufor_facewear\Data\UI\tmtm_icon_camoNet_jungle_ca.paa";
-
-		hiddenSelectionsTextures[] = {"", "\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_5_ca.paa"};
-
-        TMTM_NVG_CAMONETTING_ITEMINFO("\tmtm\pack\blufor_facewear\Data\tmtm_camoNet_jungle_5_ca.paa")
-	};
+	#include "include/CfgWeapons.hpp"
 };
