+++ addons/blufor_facewear/include/CfgGlasses.hpp
@@ -1,0 +1,1 @@
+#include "camoNetting/CfgGlasses.hpp"
