+++ addons/blufor_facewear/include/CfgWeapons.hpp
@@ -1,0 +1,2 @@
+#include "balaclava/CfgWeapons.hpp"
+#include "camoNetting/CfgWeapons.hpp"
